refactor: 수정을 위한 데이터 세팅하는 코드 모듈화
